chore(conversation): Move conversation/auth and conversation/without_content to ConversationResource #2528
