+++ front/lib/workspace_usage.ts
@@ -10,7 +10,6 @@ import {
   UserMessage,
 } from "@app/lib/models/assistant/conversation";
 import { AgentMessageFeedbackResource } from "@app/lib/resources/agent_message_feedback_resource";
-import { ConversationResource } from "@app/lib/resources/conversation_resource";
 import { DataSourceResource } from "@app/lib/resources/data_source_resource";
 import { getFrontReplicaDbConnection } from "@app/lib/resources/storage";
 import { UserModel } from "@app/lib/resources/storage/models/user";
@@ -552,8 +551,11 @@ export async function checkWorkspaceActivity(auth: Authenticator) {
     where: { workspaceId: auth.getNonNullableWorkspace().id },
   });
 
-  const hasRecentConversation = await ConversationResource.listAll(auth, {
+  // INFO: keep accessing the model for now to avoid circular deps warning
+  const owner = auth.getNonNullableWorkspace();
+  const hasRecentConversation = await Conversation.findAll({
     where: {
+      workspaceId: owner.id,
       updatedAt: { [Op.gte]: sevenDaysAgo },
     },
   });
